refactor: html rendering into new class
